pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
google-api-python-client = "^2.78.0"
google-auth-oauthlib = "^1.0.0"
xds-protos = "^0.0.11"
[tool.poetry.dev-dependencies]
